added services

+++ src/main/java/org/example/dvdrental/repositories/UserRentalRepository.java
@@ -3,5 +3,5 @@ package org.example.dvdrental.repositories;
 import org.example.dvdrental.models.UserRental;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRentalRepository extends JpaRepository<UserRental, Integer> {
+public interface UserRentalRepository extends JpaRepository<UserRental, Long> {
 }
